// src/models/mlb/team.mlb.ts
export interface TeamMlb {
  venue_short: string;
  sport_id: string;
  league_abbrev: string;
  team_id: string;
  spring_league_id: string;
  active_sw: string;
  division: string;
  mlb_org_brief: string;
  season: string;
  first_year_of_play: string;
  state: string;
  name_short: string;
  bis_team_code: string;
  venue_id: string;
  name_display_short: string;
  name_display_long: string;
  name_display_brief: string;
  sport_code_name: string;
  spring_league: string;
  league: string;
  division_id: string;
  sport_code: string;
  time_zone_num: string;
  mlb_org: string;
  name_display_full: string;
  all_star_sw: string;
  division_abbrev: string;
  name: string;
  home_opener: string;
  phone_number: string;
  address_zip: string;
  time_zone_text: string;
  venue_name: string;
  division_full: string;
  franchise_code: string;
  city: string;
  time_zone_alt: string;
  address_state: string;
  name_abbrev: string;
  store_url: string;
  file_code: string;
  address_line3: string;
  address_line2: string;
  address_province: string;
  mlb_org_id: string;
  address_line1: string;
  spring_league_full: string;
  spring_league_abbrev: string;
  last_year_of_play: string;
  address: string;
  league_full: string;
  address_country: string;
  base_url: string;
  time_zone: string;
  address_city: string;
  team_code: string;
  mlb_org_abbrev: string;
  address_intl: string;
  time_zone_generic: string;
  website_url: string;
  sport_code_display: string;
  home_opener_time: string;
  mlb_org_short: string;
  league_id: string;
}