// src/models/mlb/player.mlb.ts
export interface PlayerMlb {
  college: string;
  end_date: string;
  pro_debut_date: string;
  status_code: string;
  name_full: string;
  name_display_first_last: string;
  birth_date: string;
  height_inches: string;
  team_id: string;
  name_last: string;
  bats: string;
  player_id: string;
  position_txt: string;
  primary_position: string;
  jersey_number: string;
  starter_sw: string;
  start_date: string;
  name_display_last_first: string;
  name_first: string;
  name_use: string;
  weight: string;
  throws: string;
  team_name: string;
  team_code: string;
  team_abbrev: string;
  height_feet: string;
}